backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
dj-database-url==3.1.0
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0


django-cors-headers==4.1.0
Pillow==10.0.0
requests==2.31.0
SpeechRecognition==3.10.0
pydub==0.25.1
numpy
pandas==2.0.3
python-dateutil==2.8.2
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.6  # For PostgreSQL
drf-yasg==1.21.5  # For API documentation